app2/requirements.txt
requests # Keep if still needed, otherwise remove
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-flask